+++ src/main/java/domain/game/ChessBoardGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.piece.*;
