+++ src/test/java/domain/game/ChessBoardGeneratorTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.piece.*;
 import org.junit.jupiter.api.BeforeAll;
